[Refactor]: added endpoints folders structure
