+++ .changeset/sweet-wolves-tan.md
@@ -1,0 +1,9 @@
+---
+'@chainlink/google-weather-adapter': patch
+'@chainlink/market-closure-adapter': patch
+'@chainlink/medianizer-adapter': patch
+'@chainlink/outlier-detection-adapter': patch
+'@chainlink/proof-of-reserves-adapter': patch
+---
+
+refactor, added endpoints folder structure
