+++ packages/composites/google-weather/src/endpoint/geojson-validation.d.ts
@@ -1,0 +1,3 @@
+declare module 'geojson-validation'
+
+declare function valid(geoJSONObject: GeoJSON, trace = false): boolean
