+++ packages/composites/google-weather/src/endpoint/index.ts
@@ -1,0 +1,1 @@
+export * as weather from './weather'
