+++ packages/composites/medianizer/src/config/index.ts
@@ -2,7 +2,11 @@ import { Requester } from '@chainlink/ea-bootstrap'
 import { Config } from '@chainlink/types'
 
 export const NAME = 'MEDIANIZER'
+export const DEFAULT_ENDPOINT = 'medianizer'
 
-export const makeConfig = (prefix = ''): Config => {
-  return Requester.getDefaultConfig(prefix)
+export const makeConfig = (prefix?: string): Config => {
+  return {
+    ...Requester.getDefaultConfig(prefix),
+    defaultEndpoint: DEFAULT_ENDPOINT,
+  }
 }
