+++ packages/composites/outlier-detection/schemas/env.json
@@ -1,6 +1,7 @@
 {
   "$id": "https://external-adapters.chainlinklabs.com/schemas/outlier-detection-adapter.json",
-  "title": "@chainlink/outlier-detection-adapter env var schema",
+  "title": "Chainlink Outlier Detection composite adapter",
+  "description": "The outlier detection composite adapter will determine if a median value from a set of source data providers is an outlier value compared to either the median value from a set of data providers, or the current on-chain value. If the deviation between these values is greater than the thresholds defined, the adapter returns the current on-chain value. If all checks passes, the adapter returns the median from the source data providers.",
   "required": [],
   "type": "object",
   "properties": {
