+++ packages/composites/outlier-detection/src/endpoint/index.ts
@@ -1,0 +1,1 @@
+export * as outlier from './outlier'
