+++ packages/composites/proof-of-reserves/src/adapter.ts
@@ -1,69 +1,20 @@
-import { Logger, util } from '@chainlink/ea-bootstrap'
+import { Builder } from '@chainlink/ea-bootstrap'
 import {
-  AdapterRequest,
-  ExecuteWithConfig,
   Config,
+  AdapterRequest,
+  APIEndpoint,
   ExecuteFactory,
-  Execute,
-  AdapterResponse,
-  AdapterContext,
+  ExecuteWithConfig,
 } from '@chainlink/types'
-import { Validator, Requester } from '@chainlink/ea-bootstrap'
-import { makeConfig, makeOptions, DEFAULT_CONFIRMATIONS } from './config'
-import { runProtocolAdapter } from './protocol'
-import { Indexer, runBalanceAdapter } from './balance'
-import { runReduceAdapter } from './reduce'
+import { makeConfig } from './config'
+import * as endpoints from './endpoint'
 
-export const makeRequestFactory =
-  (config: Config, prefix: string): Execute =>
-  async (input: AdapterRequest) =>
-    (
-      await Requester.request({
-        ...config.api,
-        method: 'post',
-        url: util.getURL(prefix, true),
-        data: input,
-      })
-    ).data as AdapterResponse
-
-// Run, log, throw on error
-export const callAdapter = async (
-  execute: Execute,
-  context: AdapterContext,
-  input: AdapterRequest,
-  tag: string,
-): Promise<AdapterResponse> => {
-  const output = await execute(input, context)
-  Logger.debug(tag, { output })
-  return output
+export const execute: ExecuteWithConfig<Config> = async (request, context, config) => {
+  return Builder.buildSelector(request, context, config, endpoints)
 }
 
-const inputParams = {
-  protocol: true,
-  indexer: true,
-  confirmations: false,
-}
-
-export const execute: ExecuteWithConfig<Config> = async (input, context, config) => {
-  const paramOptions = makeOptions()
-  const validator = new Validator(input, inputParams, paramOptions)
-
-  const jobRunID = validator.validated.jobRunID
-  const protocol = validator.validated.data.protocol.toUpperCase()
-  const indexer: Indexer = validator.validated.data.indexer.toUpperCase()
-  const confirmations = validator.validated.data.confirmations || DEFAULT_CONFIRMATIONS
-
-  const protocolOutput = await runProtocolAdapter(jobRunID, context, protocol, input.data, config)
-  const balanceOutput = await runBalanceAdapter(
-    indexer,
-    context,
-    confirmations,
-    config,
-    protocolOutput,
-  )
-  const reduceOutput = await runReduceAdapter(indexer, context, balanceOutput)
-  return reduceOutput
-}
+export const endpointSelector = (request: AdapterRequest): APIEndpoint<Config> =>
+  Builder.selectEndpoint(request, makeConfig(), endpoints)
 
 export const makeExecute: ExecuteFactory<Config> = (config) => {
   return async (request, context) => execute(request, context, config || makeConfig())
