+++ packages/composites/proof-of-reserves/src/config/index.ts
@@ -1,13 +1,17 @@
 import { Requester, util } from '@chainlink/ea-bootstrap'
 import { Config } from '@chainlink/types'
-import { adapters as BalanceAdapters, Indexer } from '../balance'
-import { adapters as ProtocolAdapters, LIST_ADAPTER, Protocol } from '../protocol'
+import { adapters as BalanceAdapters, Indexer } from '../utils/balance'
+import { adapters as ProtocolAdapters, LIST_ADAPTER, Protocol } from '../utils/protocol'
 
 export const DEFAULT_CONFIRMATIONS = 6
 
 export const NAME = 'PROOF_OF_RESERVES'
+export const DEFAULT_ENDPOINT = 'reserves'
 
-export const makeConfig = (prefix?: string): Config => Requester.getDefaultConfig(prefix)
+export const makeConfig = (prefix?: string): Config => ({
+  ...Requester.getDefaultConfig(prefix),
+  defaultEndpoint: DEFAULT_ENDPOINT,
+})
 
 export type Options = {
   protocol: Protocol[]
