+++ packages/composites/proof-of-reserves/src/endpoint/index.ts
@@ -1,0 +1,1 @@
+export * as reserves from './reserves'
