+++ packages/composites/proof-of-reserves/src/utils/PorInputAddress.ts
@@ -1,0 +1,5 @@
+export interface PorInputAddress {
+  network: string
+  chainId: string
+  address: string
+}
